+++ packages/a11y/src/utils/constants.ts
@@ -1,0 +1,20 @@
+import { generateSelectors } from '@global/factory';
+import { A11Y_ATTRIBUTE } from 'global/constants/attributes';
+
+const ATTRIBUTES_PREFIX = `fs-${A11Y_ATTRIBUTE}`;
+
+export const AUTOFOCUS_ELEMENT_KEY = 'autofocus';
+
+export const ATTRIBUTES = {
+  element: {
+    key: `${ATTRIBUTES_PREFIX}-element`,
+    values: {
+      /**
+       * Defines an element that should autofocus when a wrapper is displayed.
+       */
+      autoFocus: AUTOFOCUS_ELEMENT_KEY,
+    },
+  },
+} as const;
+
+export const [getSelector, queryElement] = generateSelectors(ATTRIBUTES);