+++ packages/copyclip/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @finsweet/attributes-copyclip
 
+## 1.5.5
+
+### Patch Changes
+
+- Updated dependencies [7542dcdb]
+  - @global/constants@1.0.2
+  - @global/factory@1.1.3
+
+## 1.5.4
+
+### Patch Changes
+
+- Updated dependencies [bdd1a78]
+  - @global/constants@1.0.1
+  - @global/factory@1.1.2
+
 ## 1.5.3
 
 ### Patch Changes